[#3545] editor / section 11 revision
