+++ akvo/rsr/spa/app/modules/editor/reporting/action-types.js
@@ -1,6 +1,18 @@
 export default {
   EDIT_FIELD: 'PE_REPORTING_EDIT_FIELD',
-  ADD_ARRAY_ITEM: 'PE_REPORTING_ADD_ARRAY_ITEM',
-  REMOVE_ARRAY_ITEM: 'PE_REPORTING_REMOVE_ARRAY_ITEM',
-  EDIT_ARRAY_ITEM_FIELD: 'PE_REPORTING_EDIT_ARRAY_ITEM_FIELD'
+  FLAGS: {
+    ADD: 'PE_REPORTING_ADD_FLAG',
+    REMOVE: 'PE_REPORTING_REMOVE_FLAG',
+    EDIT_FIELD: 'PE_REPORTING_EDIT_FLAG_FIELD',
+  },
+  FORECASTS: {
+    ADD: 'PE_REPORTING_ADD_FLAG',
+    REMOVE: 'PE_REPORTING_REMOVE_FLAG',
+    EDIT_FIELD: 'PE_REPORTING_EDIT_FLAG_FIELD',
+  },
+  LEGACIES: {
+    ADD: 'PE_REPORTING_ADD_LEGACY',
+    REMOVE: 'PE_REPORTING_REMOVE_LEGACY',
+    EDIT_FIELD: 'PE_REPORTING_EDIT_LEGACY_FIELD',
+  }
 }
