+++ akvo/rsr/spa/app/store/root-reducer.js
@@ -9,7 +9,7 @@ import locationsRdr from '../modules/editor/locations/reducers'
 import focusRdr from '../modules/editor/focus/reducers'
 import {linksRdr, docsRdr} from '../modules/editor/links/reducers'
 import {commentsRdr, keywordsRdr} from '../modules/editor/comments-n-keywords/reducers'
-import reportingRdr from '../modules/editor/reporting/reducers'
+import {reportingRdr, crsAddOtherFlagRdr, forecastsRdr, legaciesRdr} from '../modules/editor/reporting/reducers'
 
 const rootReducer = combineReducers({
   partnersRdr,
@@ -23,7 +23,10 @@ const rootReducer = combineReducers({
   docsRdr,
   commentsRdr,
   keywordsRdr,
-  reportingRdr
+  reportingRdr,
+  crsAddOtherFlagRdr,
+  forecastsRdr,
+  legaciesRdr
 })
 
 export default rootReducer
